guiDocker/app/requirements.txt
Flask
mysql-connector-python
selenium
monkeylearn
pandas
flask-cors